% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alcohol_CpGs.R
\docType{data}
\name{Alcohol_CpGs}
\alias{Alcohol_CpGs}
\title{McCartney Alcohol Predictor Weights}
\format{
A data frame with 450 rows (CpGs) and 2 variables:
\describe{
  \item{CpG}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Beta}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6158884/]>
}
\usage{
Alcohol_CpGs
}
\description{
The CpGs and Weights to calculate the McCartney Alcohol Predictor
}
\keyword{datasets}
