% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common_inputs.R
\name{align_pheno}
\alias{align_pheno}
\title{Align Pheno}
\usage{
align_pheno(pheno, Sample_ID)
}
\arguments{
\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
  \item \code{Sample_ID}: Character vector of sample IDs to match with `DNAm`.
}}
}
\description{
Filter and sort pheno given a list of ids. ids are guaranteed to be in pheno.
}
\examples{
\dontrun{
examplePheno$Sample_ID <- examplePheno$name
align_pheno(examplePheno, c("7786915023_R02C02", "7786915023_R02C02"))
}

}
\keyword{internal}
