% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anti.trafo.R
\name{anti.trafo}
\alias{anti.trafo}
\title{Developmental Age Transformation}
\usage{
anti.trafo(x, adult.age = 20)
}
\arguments{
\item{x}{A vector of sample ages}

\item{adult.age}{The age considered to be the cutoff for adulthood}
}
\value{
transformed age prediction
}
\description{
Developmental Age Transformation
}
