% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bocklandt_CpG.R
\docType{data}
\name{Bocklandt_CpG}
\alias{Bocklandt_CpG}
\title{Bocklandt CpG}
\format{
A vector of CpG names
}
\source{
<https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3120753/>
}
\usage{
Bocklandt_CpG
}
\description{
The CpG whose beta value is used to give the Bocklandt clock estimate
}
\keyword{datasets}
