% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCellPopAge.R
\name{calcCellPopAge}
\alias{calcCellPopAge}
\title{Calculate CellPopAge}
\usage{
calcCellPopAge(DNAm, pheno = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
\item \code{Sample_ID}: Character vector of sample IDs to match with \code{DNAm}.
}}
}
\value{
If you added the \code{pheno} input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
Calculates CellPopAge, an epigenetic clock trained to predict
cumulative population doublings from DNA methylation data.
}
\examples{
\dontrun{
calcCellPopAge(exampleBetas, examplePheno)
}
}
\references{
Lowe, D., et al. (2022). CellPopAge: an epigenetic clock for replicative
aging of cell populations.
\url{https://github.com/ucl-medical-genomics/CellPopAge-epigenetic-clock}
}
