% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoreClocks.R
\name{calcCoreClocks}
\alias{calcCoreClocks}
\title{calcCoreClocks}
\usage{
calcCoreClocks(DNAm, pheno, CpGImputation = NULL, imputation = F)
}
\arguments{
\item{DNAm}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{pheno}{The sample phenotype data (also with samples as rows) that the clock will be appended to. If you don't have this, simply initialize a dataframe with a single column of sample IDs or rownumbers you can append clock values to.}

\item{CpGImputation}{An optional namesd vector for the mean value of each CpG that will be input from another dataset if such values are missing here (from sample cleaning)}

\item{imputation}{Logical value that will allows you to perform (T)/ skip (F) imputation of mean values for missing CpGs. Warning: when imputation = F if there are missing CpGs, it will automatically ignore these CpGs during calculation, making the clock values less accurate.}
}
\value{
A dataframe that has column names of the core clocks giving you multiple clocks at once in an easy to compute function. These will be appended onto the existing pheno dataframe as defined in the inputs.
}
\description{
A function to calculate the core, most utilized epigenetic clocks in the literature.
}
\examples{
calcCoreClocks(exampleBetas, examplePheno, imputation = F)
}
