% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCausalityAge.R
\name{calcDamAge}
\alias{calcDamAge}
\title{Calculate Ying DamAge}
\usage{
calcDamAge(DNAm, pheno = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
\item \code{Sample_ID}: Character vector of sample IDs to match with \code{DNAm}.
}}
}
\value{
If you added the \code{pheno} input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
A function to calculate the DamAge damaging-aging epigenetic clock.
}
\examples{
DamAge <- calcDamAge(
  exampleBetas,
  examplePheno
)
DamAge
}
\references{
Please cite the source when using these clocks:

Ying, K., Liu, H., Tarkhov, A.E., Sadler, M.C., Lu, A.T., Moqri, M., Horvath,
S., Kutalik, Z., Shen, X., and Gladyshev, V.N., 2024.
Causality-enriched epigenetic age uncouples damage and adaptation.
Nature Aging, 4(2), pp.231-246. \url{https://doi.org/10.1038/s43587-023-00557-0}
}
