% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDunedinPoAm38.R
\name{calcDunedinPoAm38}
\alias{calcDunedinPoAm38}
\title{calcDunedinPoAm38}
\source{
\url{https://doi.org/10.7554/eLife.54870}
}
\usage{
calcDunedinPoAm38(DNAm, pheno = NULL, CpGImputation = NULL, imputation = F)
}
\arguments{
\item{DNAm}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{pheno}{Optional: The sample phenotype data (also with samples as rows) that the clock will be appended to.}

\item{CpGImputation}{An optional namesd vector for the mean value of each CpG that will be input from another dataset if such values are missing here (from sample cleaning)}

\item{imputation}{Logical value that will allows you to perform (T)/ skip (F) imputation of mean values for missing CpGs. Warning: when imputation = F if there are missing CpGs, it will automatically ignore these CpGs during calculation, making the clock values less accurate.}
}
\value{
If you added the optional pheno input (preferred) the function appends a column with the clock calculation and returns the dataframe. Otherwise, it will return a vector of calculated clock values in order of the
}
\description{
A function to calculate the Dunedin Pace of Aging Measure (DunedinPoAm38). This accesses the library available at github.com/danbelsky/DunedinPoAm38
}
\examples{
calcDunedinPoAm38(exampleBetas, examplePheno, imputation = F)
}
