% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrimAgeV1.R
\name{calcGrimAgeV1}
\alias{calcGrimAgeV1}
\title{calcGrimAgeV1}
\usage{
calcGrimAgeV1(DNAm, pheno)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
  \item \code{Sample_ID}: Character vector of sample IDs to match with `DNAm`.
  \item \code{Female}: Numeric vector of 1 (female) or 0 (non-female).
  \item \code{Age}: Numeric vector of chronological ages (finite values).
}}
}
\value{
A data.frame of calculated clock values appended to `pheno`.
}
\description{
A function to calculate GrimAgeV1
}
\examples{
\dontrun{
GrimAgeV1 <- calcGrimAgeV1(
  exampleBetas,
  examplePheno
)
GrimAgeV1
}
}
