% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntrinClock.R
\name{calcIntrinClock}
\alias{calcIntrinClock}
\title{Calculate IntrinClock}
\usage{
calcIntrinClock(DNAm, pheno = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
\item \code{Sample_ID}: Character vector of sample IDs to match with \code{DNAm}.
}}
}
\value{
If you added the \code{pheno} input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
Calculates the IntrinClock epigenetic age estimate using a
glmnet model with developmental age transformation.
}
\examples{
\dontrun{
calcIntrinClock(exampleBetas, examplePheno)
}
}
