% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKnight.R
\name{calcKnight}
\alias{calcKnight}
\title{calcKnight}
\usage{
calcKnight(DNAm, pheno = NULL, CpGImputation = NULL, imputation = F)
}
\arguments{
\item{DNAm}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{pheno}{Optional: The sample phenotype data (also with samples as rows) that the clock will be appended to.}

\item{CpGImputation}{An optional namesd vector for the mean value of each CpG that will be input from another dataset if such values are missing here (from sample cleaning)}

\item{imputation}{Logical value that will allows you to perform (T)/ skip (F) imputation of mean values for missing CpGs. Warning: when imputation = F if there are missing CpGs, it will automatically ignore these CpGs during calculation, making the clock values less accurate.}
}
\value{
If you added the optional pheno input (preferred) the function appends a column with the clock calculation and returns the dataframe. Otherwise, it will return a vector of calculated clock values in order of the
}
\description{
A function to calculate the Knight gestational age
}
\examples{
calcKnight(exampleBetas, examplePheno, imputation = F)
}
