% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPCBrainAge.R
\name{calcPCBrainAge}
\alias{calcPCBrainAge}
\title{Calculate PCBrainAge}
\usage{
calcPCBrainAge(DNAm, pheno = NULL, RData = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
\item \code{Sample_ID}: Character vector of sample IDs to match with \code{DNAm}.
}}

\item{RData}{Default to \code{NULL}, which loads from the default path.
Either a character string specifying the path to the
\code{PCBrainAge_data.qs2} file, or a list containing the pre-loaded data
from \code{\link[=load_PCBrainAge_data]{load_PCBrainAge_data()}}. See Details.}
}
\value{
If you added the \code{pheno} input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
Calculates PCBrainAge, a principal component-based epigenetic
clock trained on brain cortical tissue.
}
\details{
PCBrainAge requires the \code{PCBrainAge_data.qs2} object, which can be
downloaded using \code{\link[=download_methylCIPHER]{download_methylCIPHER()}}. Provide either the path to the
downloaded file or the object loaded with \code{\link[=load_PCBrainAge_data]{load_PCBrainAge_data()}} to the
\code{RData} parameter.
}
\examples{
\dontrun{
download_methylCIPHER(clocks = "PCBrainAge")
calcPCBrainAge(exampleBetas, examplePheno)
}
}
\references{
Thrush, K.L., et al. (2022). Aging the brain: multi-region methylation
principal component based clock in the context of Alzheimer's disease.
Aging, 14(14), 5641-5668.
\url{https://doi.org/10.18632/aging.204196}
}
