% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPhysAge.R
\name{calcPhysAge}
\alias{calcPhysAge}
\title{calcPhysAge}
\usage{
calcPhysAge(DNAm, pheno = NULL, ID = "Sample_ID")
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
\item \code{Sample_ID}: Character vector of sample IDs to match with \code{DNAm}.
}}

\item{ID}{Column in \code{pheno} that indicates the ID of each sample.}
}
\value{
If you added the \code{pheno} input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
A function to calculate DNAmPhysAge
}
\examples{
\dontrun{
PhysAge <- calcPhysAge(
  exampleBetas,
  examplePheno
)
PhysAge
}
}
\references{
Please cite the source when using this clock:

Arpawong, T.E., Hernandez, B., Potter, C., Leigh, R.J., Klopack, E.T., Hill, C., Fiorito, G., Smyth, L.J., O’Halloran, A.M., McGuinness, B., Faul, J.D., Kenny, R.A., McKnight, A.J, Crimmins, E.M., and McCrory, C, 2025. Physiological health Age (PhysAge): a novel multi-system molecular timepiece predicts health and mortality in older adults. GeroScience, pp.1-21. \url{https://doi.org/10.1007/s11357-025-01832-1}
}
