% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStochClocks.R
\name{calcStochClocks}
\alias{calcStochClocks}
\title{calcStochClocks}
\usage{
calcStochClocks(data.m, datPheno, ages.v = NULL, refM.m = NULL)
}
\arguments{
\item{data.m}{normalized DNAm data beta-valued matrix with rownames labeling CpGs and columns labeling samples.}

\item{ages.v}{Optional argument representing the chronological ages or surrogates thereof of the samples. Vector must be of same length as the number of samples.}
}
\value{
#### A list containing the following entries
}
\description{
An R-function to estimate the DNAm-Age according to one of 3 stochastic epigenetic clocks (StocH, StocZ, StocP). These are stochastic analogues of the Horvath, Zhang and Levine/PhenoAge clocks. They are made up of exactly the same CpGs as the original clocks but were built from artificial cohorts generated by a stochastic process of age-related DNAm-change accrual. Although they were built from young and old sorted monocyte samples from the MESA study, their predictive ability is largely independent of the immune cell type or if evaluated in whole blood. This is because DNAm data is standardized before application of the clock. The stochastic clocks may be a useful addition to the original clocks, as a way of assessing if specific age-accelerations of the original clocks could be driven by an underlying stochastic process or not. This could potential shed insight on the biological mechanism underlying biological processes that are linked to the phenotypes exhibit age-acceleration (or deceleration).
}
