% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSystemsAge.R
\name{calcSystemsAge}
\alias{calcSystemsAge}
\title{calcSystemsAge}
\usage{
calcSystemsAge(DNAm, pheno = NULL, imputation = F, RData = NULL)
}
\arguments{
\item{DNAm}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{pheno}{Optional: The sample phenotype data (also with samples as rows) that the clock will be appended to.}

\item{imputation}{Logical value that will allows you to perform (T)/ skip (F) imputation of mean values for missing CpGs. Warning: when imputation = F if there are missing CpGs, it will automatically ignore these CpGs during calculation, making the clock values less accurate.}

\item{RData}{Either a character string specifying the path to the methylCIPHER folder containing
the `SystemsAge_data.RData` file, or a list/environment containing the contents of
the `SystemsAge_data.RData` file.}
}
\value{
If you added the optional pheno input (preferred) the function appends a column with the clock calculation and returns the dataframe. Otherwise, it will return a vector of calculated clock values in order of the
}
\description{
A function to calculate Systems Age
}
\examples{
\dontrun{
example_SystemsAge <- calcSystemsAge(exampleBetas, examplePheno, imputation = F)
example_SystemsAge
}

}
