% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCausalityAge.R
\docType{data}
\name{CausalityAge_CpGs}
\alias{CausalityAge_CpGs}
\alias{CausAge_CpGs}
\alias{AdaptAge_CpGs}
\alias{DamAge_CpGs}
\title{Ying Causality-Enriched Epigenetic Clock Weights}
\format{
A data frame with 2 variables:
\describe{
\item{CpG}{The CpG names from Illumina Array IDs used in the current clock}
\item{Beta}{The weighted linear regression beta value of fit for each CpG in the clock.}
}

An object of class \code{data.frame} with 585 rows and 2 columns.

An object of class \code{data.frame} with 999 rows and 2 columns.

An object of class \code{data.frame} with 1089 rows and 2 columns.
}
\usage{
CausAge_CpGs

AdaptAge_CpGs

DamAge_CpGs
}
\description{
The CpGs and Weights to calculate the Ying causality-enriched
epigenetic clocks: \code{CausAge} (causal aging), \code{AdaptAge} (adaptive aging),
and \code{DamAge} (damaging aging).
}
\references{
Please cite the source when using these clocks:

Ying, K., Liu, H., Tarkhov, A.E., Sadler, M.C., Lu, A.T., Moqri, M., Horvath,
S., Kutalik, Z., Shen, X., and Gladyshev, V.N., 2024.
Causality-enriched epigenetic age uncouples damage and adaptation.
Nature Aging, 4(2), pp.231-246. \url{https://doi.org/10.1038/s43587-023-00557-0}
}
\keyword{datasets}
