% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellDRIFT_data.R
\docType{data}
\name{CellDRIFT_CpGs}
\alias{CellDRIFT_CpGs}
\title{CellDRIFT CpGs}
\format{
A data frame with 2,322 rows and 1 variable:
\describe{
\item{CpG}{CpG probe ID}
}
}
\source{
\url{https://github.com/MorganLevineLab/CellDRIFT}
}
\usage{
CellDRIFT_CpGs
}
\description{
The CpG probes required by CellDRIFT (2,322 CpGs from yellow
and tan WGCNA modules).
}
\keyword{datasets}
