% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellPopAge_data.R
\docType{data}
\name{CellPopAge_CpGs}
\alias{CellPopAge_CpGs}
\title{CellPopAge CpGs}
\format{
A data frame with 42 rows and 2 variables:
\describe{
\item{CpG}{CpG probe ID}
\item{coefficient}{Non-zero model coefficient}
}
}
\source{
\url{https://github.com/ucl-medical-genomics/CellPopAge-epigenetic-clock}
}
\usage{
CellPopAge_CpGs
}
\description{
The 42 CpG probes with non-zero coefficients in the CellPopAge clock,
used by \code{getClockProbes} to report probe coverage.
}
\keyword{datasets}
