% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common_inputs.R
\name{check_DNAm}
\alias{check_DNAm}
\title{Check DNA Methylation Data Matrix}
\usage{
check_DNAm(DNAm, missing_allowed = TRUE)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{missing_allowed}{A logical value indicating whether missing values are allowed in the matrix. Default is \code{TRUE}.}
}
\value{
Returns \code{invisible(TRUE)} if all checks pass.
}
\description{
This function performs a series of checks on the DNA methylation data matrix to ensure it is properly formatted and contains valid data.
}
\details{
The function performs the following checks and stops with an error if any of the following conditions are not met:
\itemize{
  \item \code{DNAm} is a matrix with double values, with at least one row and one column.
  \item If \code{missing_allowed} is \code{FALSE}, \code{DNAm} must not contain any missing values.
  \item If \code{missing_allowed} is \code{TRUE}, \code{DNAm} may contain missing values, but no column or row can be entirely \code{NA}.
  \item Column names of \code{DNAm} must be unique character strings.
}
Additionally, the function issues warnings for the following conditions:
\itemize{
  \item If row names of \code{DNAm} are not set, as they should typically represent sample names.
  \item If there are more samples (rows) than cpgs (column), which is unusual for DNA methylation data.
  \item No column names start with "cg", suggesting that the matrix might need to be transposed.
  \item If missing values are present and \code{missing_allowed} is \code{TRUE}, indicating that imputation will be handled by \code{impute_DNAm}.
}
}
\keyword{internal}
