% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common_inputs.R
\name{check_pheno}
\alias{check_pheno}
\title{Check Phenotype Data Frame}
\usage{
check_pheno(pheno, extra_columns = NULL)
}
\arguments{
\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include a column called `Sample_ID` that contains the ids of the samples to be matched with the beta matrix.}

\item{extra_columns}{An optional character vector specifying additional columns to check.}
}
\value{
The function performs the checks and stops with an error if any check fails.
}
\description{
This function checks the phenotype data frame (`pheno`) to ensure it contains the required columns and that the data types are correct.
}
\details{
The function performs the following checks and stops with an error if any of the following conditions are not met:
\itemize{
  \item \code{pheno} is a data frame with at least one row.
  \item \code{pheno} contains a column named "Sample_ID", which must be a character vector without missing values.
  \item If "Female" is specified in \code{extra_columns}, \code{pheno} must contain a column named "Female", which must be an integerish vector with values 0 or 1 and no missing values.
  \item If "Age" is specified in \code{extra_columns}, \code{pheno} must contain a column named "Age", which must be a numeric vector with finite values and no missing values.
}
}
\keyword{internal}
