% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeMyClocks.R
\name{citeMyClocks}
\alias{citeMyClocks}
\title{citeMyClocks}
\usage{
citeMyClocks(clockList, prettyprint = TRUE)
}
\arguments{
\item{clockList}{A character value or vector naming the function(s) you used to calculate epigenetic clocks.}

\item{prettyprint}{If you want the console output to look nice. Default is true. If you would like to store citations as a vector for export or future reference, then make FALSE.}
}
\value{
This will return a vector list of the appropriate citations printed to your console, or you can save the vector for later reference.
}
\description{
citeMyClocks
}
\examples{
citeMyClocks("calcPhenoAge")
}
