% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockOptions.R
\name{clockOptions}
\alias{clockOptions}
\title{clockOptions}
\usage{
clockOptions()
}
\value{
A list of the available clocks to calculate, including those in the prcPhenoAge and DunedinPoAm packages
}
\description{
clockOptions
}
