% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmClockCortical_CpGs.R
\docType{data}
\name{DNAmClockCortical_CpGs}
\alias{DNAmClockCortical_CpGs}
\title{DNAmClockCortical Weights}
\format{
A data frame with 347 rows (CpGs) and 2 variables:
\describe{
  \item{CpGs}{The CpG names from Illumina Array IDs used in the current clock}
  \item{coef}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://github.com/gemmashireby/CorticalClock>
}
\usage{
DNAmClockCortical_CpGs
}
\description{
The CpGs and Weights to calculate the Pediatric-Buccal-Epigenetic (PEDBE) Clock
}
\keyword{datasets}
