% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmClockCortical_imputeRef.R
\docType{data}
\name{DNAmClockCortical_imputeRef}
\alias{DNAmClockCortical_imputeRef}
\title{Imputation values for missing data in DNAmClockCortical}
\format{
A named vector of CpGs and values for reference.
}
\source{
<https://github.com/gemmashireby/CorticalClock>
}
\usage{
DNAmClockCortical_imputeRef
}
\description{
Values for imputation of missing values in the DNAmClockCortical calculation.
}
\keyword{datasets}
