% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmFI_Li_CpGs.R
\docType{data}
\name{DNAmFI_Li_CpGs}
\alias{DNAmFI_Li_CpGs}
\title{DNAmFI_Li Weights}
\format{
A data frame with 20 rows (CpGs) and 2 variables:
\describe{
\item{CpG}{The Illumina ID of the CpG included}
\item{Beta}{The coefficient weight for calculating DNAmFI_Li}
}
}
\usage{
DNAmFI_Li_CpGs
}
\description{
The regression values to calculate DNAmFI_Li
(formerly eFRS, the 20 CpG epigenetic Framingham Risk Score)
}
\keyword{datasets}
