% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmFitAge_data.R
\docType{data}
\name{DNAmFitAge_CpGs}
\alias{DNAmFitAge_CpGs}
\title{DNAmFitAge CpGs}
\format{
A data frame with 627 rows and 1 variable:
\describe{
\item{CpG}{CpG probe ID}
}
}
\source{
\url{https://github.com/kristenmcgreevy/DNAmFitAge}
}
\usage{
DNAmFitAge_CpGs
}
\description{
The CpG probes required by the DNAmFitAge clock.
}
\keyword{datasets}
