% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmIC_CpGs.R
\docType{data}
\name{DNAmIC_CpGs}
\alias{DNAmIC_CpGs}
\title{DNAmIC Weights}
\format{
A data frame with 91 rows (CpGs) and 2 variables:
\describe{
\item{CpG}{The Illumina ID of the CpG included}
\item{Beta}{The coefficient weight for calculating DNAmIC}
}
}
\usage{
DNAmIC_CpGs
}
\description{
The regression values to calculate the 91 CpG DNAmIC
(DNA methylation Inflammatory Composite)
}
\keyword{datasets}
