% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmStress_data.R
\docType{data}
\name{DNAmStress_CpGs}
\alias{DNAmStress_CpGs}
\title{DNAmStress CpGs}
\format{
A data frame with 211 rows (CpGs) and 2 variables:
\describe{
\item{CpG}{CpG probe ID}
\item{coefficient}{Model coefficient}
}
}
\source{
Katrinli et al. (2023) Biological Psychiatry, Table S10.
}
\usage{
DNAmStress_CpGs
}
\description{
CpG probes and coefficients for the DNAmStress (MS Stress)
methylation index.
}
\keyword{datasets}
