% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmStress_data.R
\docType{data}
\name{DNAmStress_ref}
\alias{DNAmStress_ref}
\title{DNAmStress Reference Data}
\format{
A list with one element:
\describe{
\item{intercept}{Numeric scalar intercept (-4.494069).}
}
}
\usage{
DNAmStress_ref
}
\description{
Intercept for the DNAmStress clock.
}
\keyword{datasets}
