% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCIPHER_path.R
\name{download_methylCIPHER}
\alias{download_methylCIPHER}
\title{Download methylCIPHER Clock Data}
\usage{
download_methylCIPHER(
  clocks = c("all", "SystemsAge", "PCClocks"),
  path = NULL,
  force = FALSE
)
}
\arguments{
\item{clocks}{A character vector specifying which clocks to download. Options
are `"all"`, `"SystemsAge"`, and `"PCClocks"`. Default is "all", which includes
both `"SystemsAge"` and `"PCClocks"`.}

\item{path}{The directory path where the downloaded files will be saved.
If \code{NULL}, it defaults to the path returned by [get_methylCIPHER_path()].}

\item{force}{A logical value indicating whether to force the download even if
the files already exist. Default is \code{FALSE}.}
}
\value{
Invisibly returns `TRUE` if downloads were attempted or if there was
nothing to download.
}
\description{
This function downloads external data for specific methylCIPHER clocks to a specified directory.
}
\details{
Some clocks have large data dependencies that has to be first downloaded
before the clocks can be calculated. This function download the requested clock
to a specified path or default path returned by [get_methylCIPHER_path()]. SEe
[get_methylCIPHER_path()] on instructions on how to set/change the default path.
}
\examples{
\dontrun{
# Download all clocks to the default path
download_methylCIPHER()

# Download only "SystemsAge" to a custom path
download_methylCIPHER(clocks = "SystemsAge", path = "/path/to/directory")

# Force re-download of specified clock
download_methylCIPHER(clocks = "SystemsAge", force = TRUE)
}

}
