% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DunedinCpGs.R
\docType{data}
\name{DunedinPACE_CpGs}
\alias{DunedinPACE_CpGs}
\title{DunedinPACE CpGs}
\format{
A character vector of 173 CpG probe IDs.
}
\source{
Extracted from DunedinPACE::getRequiredProbes()
}
\usage{
DunedinPACE_CpGs
}
\description{
The CpG probes required by the DunedinPACE pace-of-aging clock.
}
\keyword{datasets}
