% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DunedinCpGs.R
\docType{data}
\name{DunedinPoAm38_CpGs}
\alias{DunedinPoAm38_CpGs}
\title{DunedinPoAm38 CpGs}
\format{
A character vector of 46 CpG probe IDs.
}
\source{
Extracted from DunedinPoAm38::getRequiredProbes()
}
\usage{
DunedinPoAm38_CpGs
}
\description{
The CpG probes required by the DunedinPoAm38 pace-of-aging clock.
}
\keyword{datasets}
