% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFRS_CpGs.R
\docType{data}
\name{eFRS_CpGs}
\alias{eFRS_CpGs}
\title{eFRS Weights}
\format{
A data frame with 20 rows (CpGs) and 2 variables:
\describe{
\item{CpG}{The Illumina ID of the CpG included}
\item{Beta}{The coefficient weight for calculating eFRS}
}
}
\usage{
eFRS_CpGs
}
\description{
The regression values to calculate the 20 CpG eFRS
(epigenetic Framingham Risk Score)
}
\keyword{datasets}
