% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiToc2_CpGs.R
\docType{data}
\name{EpiToc2_CpGs}
\alias{EpiToc2_CpGs}
\title{EpiToc2 CpGs}
\format{
A vector of CpG names
\describe{
  \item{delta}{estimated de-novo methylation parameter}
  \item{beta}{ground-state methylation parameter}
}
}
\source{
<https://zenodo.org/record/2632938#.YfGA3S-B2Cg>
}
\usage{
EpiToc2_CpGs
}
\description{
The dataframe and CpGs necesary to calculate EpiToc2, as found from Teschendorff's original code.
}
\keyword{datasets}
