% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiToc_CpGs.R
\docType{data}
\name{EpiToc_CpGs}
\alias{EpiToc_CpGs}
\title{EpiToc CpGs}
\format{
A vector of CpG names
}
\source{
<https://zenodo.org/record/2632938#.YfGA3S-B2Cg>
}
\usage{
EpiToc_CpGs
}
\description{
The CpGs whose beta values are summed to give the EpiToc estimate
}
\keyword{datasets}
