% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleBetas.R
\docType{data}
\name{exampleBetas}
\alias{exampleBetas}
\title{Example Methylation}
\format{
A matrix with 5 rows and 1511 variables:
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE55763>
}
\usage{
exampleBetas
}
\description{
A small sample dataset of 5 individuals' methylation beta values from Lehne et al. 2015, subset to just over 1500 CpGs, including those needed for PhenoAge.
}
\keyword{datasets}
