% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePheno.R
\docType{data}
\name{examplePheno}
\alias{examplePheno}
\title{Example Pheno Data}
\format{
A matrix with 5 rows and 6 variables:
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE55763>
}
\usage{
examplePheno
}
\description{
A small sample dataset of 5 individuals' phenotype values from Lehne et al. 2015
}
\keyword{datasets}
