% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatHorvathOnline.R
\name{formatHorvathOnline}
\alias{formatHorvathOnline}
\title{formatHorvathOnline}
\usage{
formatHorvathOnline(DNAm, Pheno, writePath)
}
\arguments{
\item{DNAm}{The methylation Beta values you will calculate epigenetic clocks with, where columns are CpGs, and rows are samples.}

\item{Pheno}{The phenotype data for samples, including a numeric column named "Age" and a numeric (1/0) column named "Female". Ensure that the Pheno frame has the same sample order as DNAm before running this.}

\item{writePath}{The path to the directory where you want the properly formatted DNAm and Pheno files (.csv) saved to be uploaded.}
}
\value{
Formatted .csv files for input to the online Horvath calculator at <http://dnamage.genetics.ucla.edu/home>. This will provide access to GrimAge, Beta value based blood cell composition, and a few epigenetic clocks.
}
\description{
A formatting Function for the Horvath Online Calculator
}
\details{
This is a tool to prepare the Pheno and DNAm dataframes for the Horvath online calculator. This might be necessary if GrimAge estimation, or methylation-beta value based blood cell composition information is desired. Please note that while attempts have been made to ensure that the Pheno frame is reformatted properly, we caution you to check the output .csv, esp the Female and Age columns, as for any automatic formatter to ensure reasonable outputs.
}
\examples{
formatHorvathOnline(exampleBetas, examplePheno)
}
