% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Garagnani_CpG.R
\docType{data}
\name{Garagnani_CpG}
\alias{Garagnani_CpG}
\title{Garagnani CpG}
\format{
A vector of CpG names
}
\source{
<https://doi.org/10.1111/acel.12005>
}
\usage{
Garagnani_CpG
}
\description{
The CpG whose beta value is used to give the Garagnani clock estimate
}
\keyword{datasets}
