% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_methylCIPHER_default_path}
\alias{get_methylCIPHER_default_path}
\title{Get Default Path for methylCIPHER Data}
\usage{
get_methylCIPHER_default_path()
}
\value{
The default path to the methylCIPHER data directory.
}
\description{
This function determines the default directory path where methylCIPHER data,
required for SystemsAge and PCClocks, should be stored.
}
