% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_methylCIPHER_path}
\alias{get_methylCIPHER_path}
\title{Get Path for methylCIPHER Data}
\usage{
get_methylCIPHER_path()
}
\value{
The path to the methylCIPHER data directory.
}
\description{
This function retrieves the path to the methylCIPHER data directory, which is
required for SystemsAge and PCClocks. It first checks if a custom path has been
set using [set_methylCIPHER_path()] or [options()]. If no custom path is found, it returns
the default path obtained from [get_methylCIPHER_default_path()].
}
\examples{
get_methylCIPHER_path() # Default Path
set_methylCIPHER_path(".") # Change to "."
get_methylCIPHER_path() # New Path
}
