% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClockInfo.R
\name{getClockInfo}
\alias{getClockInfo}
\title{getClockInfo}
\usage{
getClockInfo()
}
\value{
A data frame of clock metadata including clock name, author, year,
PMID, trained phenotype, number of CpGs, training cohort, tissues,
age range, and array type.
}
\description{
getClockInfo
}
\examples{
getClockInfo()
}
