% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClockProbes.R
\name{getClockProbes}
\alias{getClockProbes}
\title{getClockProbes}
\usage{
getClockProbes(DNAm)
}
\arguments{
\item{DNAm}{The methylation Beta values you will calculate epigenetic clocks with, where columns are CpGs, and rows are samples.}
}
\value{
A table which compares the number of probes required by the full clocks in the current package, and the ones in your data. If you are missing many probes for a clock, imputation is necessary, though you might consider using a different clock or dataset if possible.
}
\description{
getClockProbes
}
\examples{
getClockProbes(exampleBetas)
}
