% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hannum_CpGs.R
\docType{data}
\name{Hannum_CpGs}
\alias{Hannum_CpGs}
\title{Hannum Weights}
\format{
A data frame with 71 rows (CpGs) and 2 variables:
\describe{
  \item{Marker}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Coefficient}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://doi.org/10.1016/j.molcel.2012.10.016>
}
\usage{
Hannum_CpGs
}
\description{
The CpGs and Weights to calculate the Hannum Clock
}
\keyword{datasets}
