% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horvath1_CpGs.R
\docType{data}
\name{Horvath1_CpGs}
\alias{Horvath1_CpGs}
\title{Horvath Multi-Tissue Weights}
\format{
A data frame with 353 rows (CpGs) and 2 variables:
\describe{
  \item{CpGmarker}{The CpG names from Illumina Array IDs used in the current clock}
  \item{CoefficientTraining}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://genomebiology.biomedcentral.com/articles/10.1186/gb-2013-14-10-r115>
}
\usage{
Horvath1_CpGs
}
\description{
The CpGs and Weights to calculate the Horvath Multi-Tissue clock
}
\keyword{datasets}
