% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horvath2_CpGs.R
\docType{data}
\name{Horvath2_CpGs}
\alias{Horvath2_CpGs}
\title{Horvath Skin & Blood Weights}
\format{
A data frame with 391 rows (CpGs) and 2 variables:
\describe{
  \item{ID}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Coef}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://pubmed.ncbi.nlm.nih.gov/30048243/>
}
\usage{
Horvath2_CpGs
}
\description{
The CpGs and Weights to calculate the Horvath Skin & Blood clock
}
\keyword{datasets}
