% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorvathOnlineRef.R
\docType{data}
\name{HorvathOnlineRef}
\alias{HorvathOnlineRef}
\title{Horvath Online Calculator Reference File}
\format{
A data frame with 30084 rows (CpGs) and 5 variables:
\describe{
  \item{Name}{Most Important Column: The CpG names from Illumina Array IDs used by the Online Calculator.}
  \item{Gene_ID}{The gene ID number from the original file.}
  \item{GenomeBuild}{The Genome Build number the reference was built from.}
  \item{Chr}{Chromosome of the CpG.}
  \item{Accession}{Gene Accession number for reference.}
}
}
\source{
<http://dnamage.genetics.ucla.edu>
}
\usage{
HorvathOnlineRef
}
\description{
The file containing the essential Probes for submission to the Horvath New Advanced Calculator Online
}
\keyword{datasets}
