% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSInCHPhenoAge_CpGs.R
\docType{data}
\name{HRSInCHPhenoAge_CpGs}
\alias{HRSInCHPhenoAge_CpGs}
\title{HRSInChPhenoAge Weights}
\format{
A data frame with 959 rows (CpGs) and 2 variables:
\describe{
  \item{CpG}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Weight}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<[get-from-Albert]>
}
\usage{
HRSInCHPhenoAge_CpGs
}
\description{
The CpGs and Weights to calculate Updated Pheno Age. Intercept is 52.8334080
}
\keyword{datasets}
