% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypoClock_CpGs.R
\docType{data}
\name{hypoClock_CpGs}
\alias{hypoClock_CpGs}
\title{HypoClock CpGs}
\format{
A vector of CpG names
}
\source{
<https://zenodo.org/record/2632938#.YfGA3S-B2Cg>
}
\usage{
hypoClock_CpGs
}
\description{
The CpGs whose beta values are summed to give the HypoClock estimate
}
\keyword{datasets}
