% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_DNAm.R
\name{impute_DNAm}
\alias{impute_DNAm}
\title{Impute Missing Values and Add Missing CpGs in DNA Methylation Matrix}
\usage{
impute_DNAm(DNAm, method = c("mean"), CpGs = NULL, subset = TRUE)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{method}{The method used to impute missing values within existing CpG sites. Currently, only "mean" is supported, which imputes missing values with the column mean.}

\item{CpGs}{A named numeric vector where names are CpGs names and values are used to impute completely missing CpG sites across all samples. If \code{NULL}, no new CpG sites are added, and only missing values in existing CpG sites are imputed.}

\item{subset}{A boolean. If \code{TRUE}, then only the CpGs provided in the \code{CpGs} argument are imputed and returned. Otherwise, all CpGs that have missing values in \code{DNAm} will be imputed and the full DNAm matrix with missing CpGs are returned.}
}
\value{
An imputed numeric matrix with no missing values
}
\description{
This function imputes missing values in a DNA methylation matrix and adds completely missing CpG sites using values provided in a named vector.
}
\details{
The function handles two types of missing data in a DNA methylation beta matrix: 1) completely missing CpGs are filled in with the provided values and 2) missing values of existing CpGs are filled in by supported method (e.g, mean imputation).
If \code{CpGs} is \code{NULL} only missing values in the existing columns are imputed.
}
\examples{
\dontrun{
# Sample DNAm matrix with missing values
DNAm <- matrix(
  c(0.1, 0.2, NA, 0.4, 0.5, 0.6),
  nrow = 2,
  dimnames = list(c("sample1", "sample2"), c("cpg1", "cpg2", "cpg3"))
)
# Named CpGs vector where names are CpGs and values are used to impute completely missing probes
CpGs <- c(cpg1 = 0.3, cpg2 = 0.3, cpg3 = 0.3, cpg4 = 0.7)
# Impute the matrix
imputed_DNAm <- impute_DNAm(DNAm, method = "mean", CpGs = CpGs)
imputed_DNAm
}
}
