% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeeRefinedRobust_CpGs.R
\docType{data}
\name{LeeRefinedRobust_CpGs}
\alias{LeeRefinedRobust_CpGs}
\title{Lee Robust Refined Placental Age}
\format{
A data frame with 395 rows (CpGs) and 2 variables:
\describe{
  \item{CpG}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Coef}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://doi.org/10.18632%2Faging.102049>
}
\usage{
LeeRefinedRobust_CpGs
}
\description{
The CpGs and Weights to calculate the Lee Robust and Refined Placental Age clock
}
\keyword{datasets}
