% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCIPHER_path.R
\name{load_data_function}
\alias{load_data_function}
\title{Dependency Factory}
\usage{
load_data_function(object_name, object_hash)
}
\arguments{
\item{object_name}{name of object ie SystemsAge_data}

\item{object_hash}{hash of object generated in data-raw}
}
\description{
Some functions like [calcPCClocks()] and [calcSystemsAge()] requires large external
data to be downloaded and loaded into R. The best way to increase the speed
of this process is to use qs2 storage format. This function factory generate
functions to load in these data based on a template.
}
\keyword{internal}
