% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCIPHER_path.R
\name{load_PCBrainAge_data}
\alias{load_PCBrainAge_data}
\title{Load PCBrainAge Data}
\usage{
load_PCBrainAge_data(path = NULL)
}
\arguments{
\item{path}{Character string specifying the file path to the .qs2 file.
If \code{NULL}, loads from the path at \code{\link[=get_methylCIPHER_path]{get_methylCIPHER_path()}}.}
}
\value{
A list containing objects needed to calculate PCBrainAge.
}
\description{
A wrapper that loads PCBrainAge data from a specified or default location.
}
\examples{
\dontrun{
PCBrainAge_data <- load_PCBrainAge_data()
PCBrainAge_data <- load_PCBrainAge_data("path/to/PCBrainAge_data.qs2")
}
}
