% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCIPHER_path.R
\name{load_PCClocks_data}
\alias{load_PCClocks_data}
\title{Load PCClocks Data}
\usage{
load_PCClocks_data(path = NULL)
}
\arguments{
\item{path}{Character string specifying the file path to the .qs2 file.
If `NULL`, loads from the path at [get_methylCIPHER_path()].}
}
\value{
A list containing objects needed to calculate PCClocks.
}
\description{
A wrapper that loads PCClocks data from a specified or default location.
}
\examples{
\dontrun{
PCClocks_data <- load_PCClocks_data()
PCClocks_data <- load_PCClocks_data("path/to/PCClocks_data.qs2")
}
}
