% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCIPHER_path.R
\name{load_SystemsAge_data}
\alias{load_SystemsAge_data}
\title{Load SystemsAge Data}
\usage{
load_SystemsAge_data(path = NULL)
}
\arguments{
\item{path}{Character string specifying the file path to the .qs2 file.
If `NULL`, loads from the path at [get_methylCIPHER_path()].}
}
\value{
A list containing objects needed to calculate SystemsAge.
}
\description{
A wrapper that loads SystemsAge data from a specified or default location.
}
\examples{
\dontrun{
SystemsAge_data <- load_SystemsAge_data()
SystemsAge_data <- load_SystemsAge_data("path/to/SystemsAge_data.qs2")
}
}
