% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simple_clock.R
\name{make_simple_clock}
\alias{make_simple_clock}
\title{Create a simple linear epigenetic clock calculator}
\usage{
make_simple_clock(obj, intercept, clock, cpg, coefficients)
}
\arguments{
\item{obj}{List/data.frame with CpGs and coefficients}

\item{intercept}{Numeric intercept}

\item{clock}{Column name for the resulting clock}

\item{cpg}{Name of CpG vector element in \code{obj}}

\item{coefficients}{Name of coefficient vector element in \code{obj}}
}
\value{
Function that adds the clock to a phenotype data frame
}
\description{
Create a simple linear epigenetic clock calculator
}
\keyword{internal}
