% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mayne_CpGs.R
\docType{data}
\name{Mayne_CpGs}
\alias{Mayne_CpGs}
\title{Mayne Gestational Age CpGs}
\format{
A data frame with 62 rows (CpGs) and 2 variables:
\describe{
  \item{CpG}{The CpG names from Illumina Array IDs used in the current clock}
  \item{coef}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://doi.org/10.2217/epi-2016-0103>
}
\usage{
Mayne_CpGs
}
\description{
The CpGs and Weights to calculate the Mayne Gestational Age predictor
}
\keyword{datasets}
