% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mayne_impute.R
\docType{data}
\name{Mayne_impute}
\alias{Mayne_impute}
\title{Mayne Imputation Values}
\format{
A vector of length 18437 providing the gold standard means from a large dataset for gestational age investigation.
}
\source{
<https://doi.org/10.2217/epi-2016-0103>
}
\usage{
Mayne_impute
}
\description{
The CpGs and gold standard means for imputation as reported by Mayne et. al for calcMayne
}
\keyword{datasets}
