% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanimpute.R
\name{mean_impute}
\alias{mean_impute}
\title{Mean imputation for Beta Matrix}
\usage{
mean_impute(DNAm)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}
}
\value{
A matrix of methylation beta values with missing values imputed using column means.
}
\description{
Performs column mean imputation for missing values in a methylation beta matrix, where rows represent samples and columns represent CpGs.
}
\keyword{internal}
