% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanimpute.R
\name{meanimpute}
\alias{meanimpute}
\title{Mean Imputation for singly missing CpGs}
\usage{
meanimpute(x)
}
\arguments{
\item{x}{A dataframe of CpG Betas with missing values}
}
\value{
Mean Imputed dataframe
}
\description{
Mean Imputation for singly missing CpGs
}
