% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiAge_CpGs.R
\docType{data}
\name{MiAge_CpGs}
\alias{MiAge_CpGs}
\title{MiAge Weights}
\format{
A data frame with 268 rows (CpGs) and 2 variables:
\describe{
  \item{CpGs}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Age-hyper/Age-hypo}{An indication of whether the CpG becomes hyper or hypo methylated with age}
}
}
\source{
<http://www.columbia.edu/~sw2206/softwares.htm>
}
\usage{
MiAge_CpGs
}
\description{
The CpGs and Weights to calculate Updated Pheno Age. Intercept is 52.8334080
}
\keyword{datasets}
