% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiAge_fr2.R
\name{MiAge_fr2}
\alias{MiAge_fr2}
\title{MiAge fr2}
\usage{
MiAge_fr2(x, b, c, d, betaj)
}
\value{
objective fuction summed over all CpG istes for patient j
}
\description{
#' @description A function used by the original MiAge code found <http://www.columbia.edu/~sw2206/softwares.htm>
}
