% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiAge_grr2.R
\name{MiAge_grr2}
\alias{MiAge_grr2}
\title{MiAge grr2}
\usage{
MiAge_grr2(x, b, c, d, betaj)
}
\value{
The derivative of fr2 with respect to n_{j}
}
\description{
#' @description A function used by the original MiAge code found <http://www.columbia.edu/~sw2206/softwares.htm>
}
