% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiAge_mitotic.age.R
\name{MiAge_mitotic.age}
\alias{MiAge_mitotic.age}
\title{MiAge Mitotic Age}
\usage{
MiAge_mitotic.age(beta, b, c, d)
}
\value{
The MiAge estimation
}
\description{
A function used by the original MiAge code found <http://www.columbia.edu/~sw2206/softwares.htm>
}
