% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiAge_parameters.R
\docType{data}
\name{MiAge_parameters}
\alias{MiAge_parameters}
\title{MiAge Parameters}
\format{
A list object of length 4 containing the parameters necessary to run MiAge_mitotic.age
}
\source{
<http://www.columbia.edu/~sw2206/softwares.htm>
}
\usage{
MiAge_parameters
}
\description{
The CpGs and Weights to calculate Updated Pheno Age. Intercept is 52.8334080
}
\keyword{datasets}
