% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{param_template_female_age}
\alias{param_template_female_age}
\title{Inherit parameters for functions. With Female and Age}
\usage{
param_template_female_age(DNAm, pheno, imputation)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
  \item \code{Sample_ID}: Character vector of sample IDs to match with the beta matrix.
  \item \code{Female}: Numeric vector of 1 (female) or 0 (non-female).
  \item \code{Age}: Numeric vector of chronological ages (finite values).
}}

\item{imputation}{Logical value that will allows you to perform (T)/ skip (F) imputation of mean values for missing CpGs. Warning: when imputation = F if there are missing CpGs, it will automatically ignore these CpGs during calculation, making the clock values less accurate.}
}
\description{
Template function to inherit parameters for other functions requiring phenotype data.
}
\keyword{internal}
