% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCBrainAge_CpGs.R
\docType{data}
\name{PCBrainAge_CpGs}
\alias{PCBrainAge_CpGs}
\title{PCBrainAge CpGs}
\format{
A character vector of CpG probe IDs.
}
\source{
\url{https://github.com/MorganLevineLab/calcPCBrainAge}
}
\usage{
PCBrainAge_CpGs
}
\description{
The CpG probes required by the PCBrainAge clock (357,852 CpGs).
}
\keyword{datasets}
