% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEDBE_CpGs.R
\docType{data}
\name{PEDBE_CpGs}
\alias{PEDBE_CpGs}
\title{PEDBE Weights}
\format{
A data frame with 94 rows (CpGs) and 2 variables:
\describe{
  \item{ID}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Coef}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://github.com/kobor-lab/Public-Scripts>
}
\usage{
PEDBE_CpGs
}
\description{
The CpGs and Weights to calculate the Pediatric-Buccal-Epigenetic (PEDBE) Clock
}
\keyword{datasets}
