% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoAge_CpGs.R
\docType{data}
\name{PhenoAge_CpGs}
\alias{PhenoAge_CpGs}
\title{PhenoAge Weights}
\format{
A data frame with 513 rows (CpGs) and 9 variables:
\describe{
  \item{CpG:}{The Illumina ID of the CpG included}
  \item{Weight:}{The weight when performing weighted linear regression for calculating PhenoAge}
}
}
\source{
<https://dx.doi.org/10.18632%2Faging.101414>
}
\usage{
PhenoAge_CpGs
}
\description{
The regression values to calculate PhenoAge
}
\keyword{datasets}
