% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysAge_CpGs.R
\docType{data}
\name{PhysAge_CpGs}
\alias{PhysAge_CpGs}
\title{PhysAge Weights}
\format{
A data frame with 1711 rows (CpGs) and 2 variables:
\describe{
\item{cpg}{The CpG names from Illumina Array IDs used in the current clock}
\item{mean}{These are mean beta values from the HRS training set}
}
}
\source{
\url{https://github.com/em-arpy/DNAm_PhysAge}
}
\usage{
PhysAge_CpGs
}
\description{
The CpGs and Weights to calculate the DNA methylation (DNAm) Physiological health Age (PhysAge) Clock
}
\keyword{datasets}
