% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeNAcol}
\alias{removeNAcol}
\title{Remove Columns with All NA Values from DNAm}
\usage{
removeNAcol(DNAm)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}
}
\value{
DNAm with columns removed where all values are NA.
}
\description{
Remove columns from the DNAm matrix where all values are NA.
}
\keyword{internal}
