% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNAprobes.R
\name{removeNAprobes}
\alias{removeNAprobes}
\title{Remove All-NA CpG Probes}
\usage{
removeNAprobes(DNAm)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and
CpGs as columns.}
}
\value{
The DNAm matrix with all-NA columns removed.
}
\description{
Removes CpG columns where all samples are NA.
}
\examples{
\dontrun{
DNAm <- removeNAprobes(DNAm)
}
}
