% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepliTali_data.R
\docType{data}
\name{RepliTali_CpGs}
\alias{RepliTali_CpGs}
\title{RepliTali CpGs}
\format{
A data frame with 87 rows (CpGs) and 2 variables:
\describe{
\item{CpG}{CpG probe ID}
\item{coefficient}{Model coefficient}
}
}
\usage{
RepliTali_CpGs
}
\description{
CpG probes and coefficients for the RepliTali mitotic age clock.
}
\keyword{datasets}
