% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepliTali_data.R
\docType{data}
\name{RepliTali_ref}
\alias{RepliTali_ref}
\title{RepliTali Reference Data}
\format{
A list with one element:
\describe{
\item{intercept}{Numeric scalar intercept (101.5896).}
}
}
\usage{
RepliTali_ref
}
\description{
Intercept for the RepliTali clock.
}
\keyword{datasets}
