% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepliTali_data.R
\docType{data}
\name{RepliTaliNorm_CpGs}
\alias{RepliTaliNorm_CpGs}
\title{RepliTaliNorm CpGs}
\format{
A data frame with 218 rows (CpGs) and 2 variables:
\describe{
\item{CpG}{CpG probe ID}
\item{coefficient}{Model coefficient}
}
}
\usage{
RepliTaliNorm_CpGs
}
\description{
CpG probes and coefficients for the RepliTaliNorm normalized
mitotic age clock.
}
\keyword{datasets}
