% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepliTali_data.R
\docType{data}
\name{RepliTaliNorm_ref}
\alias{RepliTaliNorm_ref}
\title{RepliTaliNorm Reference Data}
\format{
A list with one element:
\describe{
\item{intercept}{Numeric scalar intercept (75.88006).}
}
}
\usage{
RepliTaliNorm_ref
}
\description{
Intercept for the RepliTaliNorm clock.
}
\keyword{datasets}
