% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetroAgeEPICv2_CpGs.R
\docType{data}
\name{RetroAgeEPICv2_CpGs}
\alias{RetroAgeEPICv2_CpGs}
\title{RetroAgeEPICv2 Weights}
\format{
A data frame with 1378 rows (CpGs) and 2 variables:
\describe{
\item{name}{The Illumina ID of the CpG included}
\item{coefficient}{The coefficient weight for calculating RetroAgeEPICv2}
}
}
\usage{
RetroAgeEPICv2_CpGs
}
\description{
The CpGs and coefficients for calculating RetroAgeEPICv2
}
\keyword{datasets}
