% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetroAgeEPICv2_CpGs.R
\docType{data}
\name{RetroAgeEPICv2_ref}
\alias{RetroAgeEPICv2_ref}
\title{RetroAgeEPICv2 Reference Data}
\format{
A list with two elements:
\describe{
\item{intercept}{Numeric scalar intercept for the linear model.}
\item{imputation}{Named numeric vector for imputing missing CpGs.}
}
}
\usage{
RetroAgeEPICv2_ref
}
\description{
Intercept and imputation reference for RetroAgeEPICv2
}
\keyword{datasets}
