% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_methylCIPHER_path}
\alias{set_methylCIPHER_path}
\title{Set Custom Path for methylCIPHER Data}
\usage{
set_methylCIPHER_path(path)
}
\arguments{
\item{path}{A string specifying the custom path to the methylCIPHER data directory.}
}
\value{
None
}
\description{
This function allows the user to set a custom directory path for storing
methylCIPHER data, which is required for SystemsAge and PCClocks. The specified
path is stored in R's options under "methylCIPHER.path".
}
\examples{
get_methylCIPHER_path() # Default Path
set_methylCIPHER_path(".") # Change to "."
get_methylCIPHER_path() # New Path
}
