% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCG.R
\name{subsetCG}
\alias{subsetCG}
\title{subsetCG}
\usage{
subsetCG(dat, cgSet)
}
\arguments{
\item{dat}{The methylation Beta values you will need to subset, where columns are CpGs (and are named), and rows are samples.}

\item{cgSet}{The character vector of Illumina CpG probe IDs that you will be subsetting to.}
}
\value{
A new beta methylation matrix with columns of only the CpGs that you wanted to subset to.
}
\description{
A function to quickly subset the CpGs you want to work with in a methylation dataframe
}
\examples{
subsetCG(exampleBetas, HorvathOnlineRef$Name)
}
