% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo.R
\name{trafo}
\alias{trafo}
\title{trafo}
\usage{
trafo(x, adult.age = 20)
}
\arguments{
\item{x}{A vector of sample ages}

\item{adult.age}{Age boundary of adulthood--set to 20}
}
\value{
a vector of transformed ages
}
\description{
trafo
}
