% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VidalBralo_CpGs.R
\docType{data}
\name{VidalBralo_CpGs}
\alias{VidalBralo_CpGs}
\title{Vidal-Bralo Weights}
\format{
A data frame with 8 rows (CpGs) and 2 variables:
\describe{
  \item{Marker}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Coef}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4943959/>
}
\usage{
VidalBralo_CpGs
}
\description{
The CpGs and Weights to calculate the Vidal-Bralo clock
}
\keyword{datasets}
