% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weidner_CpGs.R
\docType{data}
\name{Weidner_CpGs}
\alias{Weidner_CpGs}
\title{Weidner CpGs}
\format{
A vector of CpG names
}
\source{
<https://doi.org/10.1186/gb-2014-15-2-r24>
}
\usage{
Weidner_CpGs
}
\description{
The CpGs whose beta values are summed to give the Weidner clock estimate
}
\keyword{datasets}
