% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero_cpgs}
\alias{zero_cpgs}
\title{Convert to Vector of Zeros with CpG Names}
\usage{
zero_cpgs(CpGs)
}
\value{
A numeric vector of zeros with names set to the input CpGs.
}
\description{
Convert a vector of CpGs to a numeric vector of zeros, with the names of the vector set to the CpGs.
}
\keyword{internal}
