% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zhang_10_CpG.R
\docType{data}
\name{Zhang_10_CpG}
\alias{Zhang_10_CpG}
\title{Zhang 10 CpG clock Weights}
\format{
A data frame with 10 rows (CpGs) and 2 variables:
\describe{
  \item{Marker}{The CpG names from Illumina Array IDs used in the current clock}
  \item{Coef}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
<https://www.nature.com/articles/ncomms14617>
}
\usage{
Zhang_10_CpG
}
\description{
The CpGs and Weights to calculate teh 2017 Yan Zhang 10 CpG clock
}
\keyword{datasets}
