% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zhang_CpGs.R
\docType{data}
\name{Zhang_CpGs}
\alias{Zhang_CpGs}
\title{Zhang 10 CpG Clock Weights}
\format{
A data frame with 10 rows (CpGs) and 2 variables:
\describe{
\item{Marker}{The CpG names from Illumina Array IDs used in the current clock}
\item{coef}{The weighted linear regression beta value of fit for each CpG in the clock.}
}
}
\source{
\url{https://www.nature.com/articles/ncomms14617}
}
\usage{
Zhang_CpGs
}
\description{
The CpGs and Weights to calculate the 2017 Yan Zhang 10 CpG clock
}
\keyword{datasets}
