% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{align_pheno}
\alias{align_pheno}
\title{Align Pheno}
\usage{
align_pheno(pheno, Sample_ID, ID)
}
\arguments{
\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
\item \code{Sample_ID}: Character vector of sample IDs to match with \code{DNAm}.
}}

\item{ID}{Column in \code{pheno} that indicates the ID of each sample.}
}
\description{
Filter and sort pheno given a list of ids. ids must be in pheno.
}
\examples{
\dontrun{
align_pheno(examplePheno, c("7786915023_R02C02", "7786915023_R02C02"))
}

}
\keyword{internal}
