% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrimAgeV2.R
\name{calcGrimAgeV2}
\alias{calcGrimAgeV2}
\title{calcGrimAge2V2}
\usage{
calcGrimAgeV2(datMeth, datPheno)
}
\arguments{
\item{DNAm}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{pheno}{Optional: The sample phenotype data (also with samples as rows) that the clock will be appended to.}
}
\value{
If you added the optional pheno input (preferred) the function appends a column with the clock calculation and returns the dataframe. Otherwise, it will return a vector of calculated clock values in order of the
}
\description{
A function to calculate GrimAgeV2
}
\examples{
calcPhenoAge(exampleBetas, examplePheno)


}
