% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntrinClock.R
\name{calcIntrinClock}
\alias{calcIntrinClock}
\title{Intrin Clock}
\usage{
calcIntrinClock(datMeth, datPheno)
}
\arguments{
\item{DNAm}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{pheno}{The sample phenotype data (also with samples as rows) that the clock will be appended to.}
}
\value{
If you added the optional pheno input (preferred) the function appends a column with the clock calculation and returns the dataframe. Otherwise, it will return a vector of calculated clock values in order of the
}
\description{
A function to calculate IntrinClock
}
\examples{
calcEpiDish(exampleBetas, examplePheno)

}
