% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPCClocks.R
\name{calcPCClocks}
\alias{calcPCClocks}
\title{calcPCClocks}
\usage{
calcPCClocks(DNAm, pheno, RData = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing the phenotype of the samples where each row must correspond to each row of `DNAm`. Must contains the following columns: `Female` = 1 or 0 and chronological age in years.}

\item{RData}{Either a character string specifying the path to the methylCIPHER folder containing
the `CalcAllPCClocks.RData` file, or a list/environment containing the contents of
the `CalcAllPCClocks.RData` file.}
}
\value{
A data.frame of calculated clock values appened to `pheno`.
}
\description{
A function to calculate Systems Age
}
\examples{
\dontrun{
example_PCClocks <- calcPCClocks(
  exampleBetas,
  examplePheno,
  # If `PCClocks_data.RData` has been downloaded to default folder at $HOME
  RData = get_methylCIPHER_path()
)
example_PCClocks
}

}
