% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPCClocks.R
\name{calcPCClocks}
\alias{calcPCClocks}
\title{calcPCClocks}
\usage{
calcPCClocks(DNAm, pheno, RData = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
  \item \code{Sample_ID}: Character vector of sample IDs to match with `DNAm`.
  \item \code{Female}: Numeric vector of 1 (female) or 0 (non-female).
  \item \code{Age}: Numeric vector of chronological ages (finite values).
}}

\item{RData}{Either a character string specifying the path to the methylCIPHER folder containing
the `PCClocks_data.qs2` file or a list containing the contents of
the `PCClocks_data.qs2` file created by [load_PCClocks_data()].}
}
\value{
A data.frame of calculated clock values appended to `pheno`.
}
\description{
A function to calculate PC Clocks
}
\examples{
\dontrun{
example_PCClocks <- calcPCClocks(
  exampleBetas,
  examplePheno,
  RData = get_methylCIPHER_path()
)
example_PCClocks
}
}
