% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSystemsAge.R
\name{calcSystemsAge}
\alias{calcSystemsAge}
\title{calcSystemsAge}
\usage{
calcSystemsAge(DNAm, pheno = NULL, RData = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{Optional: A data frame or matrix containing sample phenotype data, with samples as rows. The clock results will be appended to this data. Defaults to \code{NULL}.}

\item{RData}{Either a character string specifying the path to the methylCIPHER folder containing
the `SystemsAge_data.RData` file, or a list/environment containing the contents of
the `SystemsAge_data.RData` file.}
}
\value{
If you added the optional pheno input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
A function to calculate Systems Age
}
\examples{
\dontrun{
example_SystemsAge <- calcSystemsAge(
  exampleBetas,
  examplePheno,
  # If `SystemsAge_data.RData` has been downloaded to default folder at $HOME
  RData = get_methylCIPHER_path()
)
example_SystemsAge
}

}
