% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSystemsAge.R
\name{calcSystemsAge}
\alias{calcSystemsAge}
\title{calcSystemsAge}
\usage{
calcSystemsAge(DNAm, pheno = NULL, RData = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
  \item \code{Sample_ID}: Character vector of sample IDs to match with `DNAm`.
}}

\item{RData}{Either a character string specifying the path to the methylCIPHER folder containing
the `SystemsAge_data.qs2` file or a list containing the contents of
the `SystemsAge_data.qs2` file created by [load_SystemsAge_data()].}
}
\value{
If you added the `pheno` input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
A function to calculate Systems Age
}
\examples{
\dontrun{
example_SystemsAge <- calcSystemsAge(
  exampleBetas,
  examplePheno,
  RData = get_methylCIPHER_path()
)
example_SystemsAge
}

}
