% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSystemsAge.R
\name{calcSystemsAge}
\alias{calcSystemsAge}
\title{Calculate Systems Age}
\usage{
calcSystemsAge(DNAm, pheno = NULL, ID = "Sample_ID", RData = NULL)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
  \item \code{Sample_ID}: Character vector of sample IDs to match with `DNAm`.
}}

\item{ID}{Column in `pheno` that indicates the ID of each sample.}

\item{RData}{Default to \code{NULL}, which is the default path to the downloaded data folder.
Either a character string specifying the path to the folder containing
the `SystemsAge_data.qs2` file, or a list containing the contents of the
`SystemsAge_data.qs2` file loaded via [load_SystemsAge_data()]. See Details.}
}
\value{
If you added the `pheno` input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
Calculates Systems Age
}
\details{
Systems Age calculation requires the `SystemsAge_data.qs2` object, which can be
downloaded using [download_methylCIPHER()]. Provide either the path to the
downloaded `SystemsAge_data.qs2` file or the object loaded with
[load_SystemsAge_data()] to the `RData` parameter. The advantage of passing
in the loaded data is that if this function were to be run multiple times,
then the object won't be loaded in each time.
}
\examples{
\dontrun{
# Download the external data
download_methylCIPHER(clocks = "SystemsAge")

# Either path to the data
RData <- get_methylCIPHER_path()
# Or load in the object
RData <- load_SystemsAge_data(get_methylCIPHER_path())

# Calculate Systems Age using example data
result <- calcSystemsAge(
  betas = exampleBetas,
  pheno = examplePheno,
  RData = RData
)
result
}
}
