% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCIPHER_path.R
\name{get_methylCIPHER_path}
\alias{get_methylCIPHER_path}
\title{Get Path for methylCIPHER Data}
\usage{
get_methylCIPHER_path()
}
\value{
A character string specifying the path to the methylCIPHER data directory.
}
\description{
Retrieves the path to the methylCIPHER data directory. The function checks for a custom path set via
\code{\link[=set_methylCIPHER_path]{set_methylCIPHER_path()}} or the \code{methylCIPHER.path} option
(see \code{\link[=options]{options()}}). If no custom path is specified, it returns the
default path from \code{\link[=get_methylCIPHER_default_path]{get_methylCIPHER_default_path()}}.
}
\details{
Tips: To set a custom path permanently, add
\code{options(methylCIPHER.path = "non_default_path")} to your \code{.Rprofile}
file.
}
\examples{
# Returns the default path
get_methylCIPHER_path()
# Sets path to current directory
set_methylCIPHER_path(".")
# Returns the new path
get_methylCIPHER_path()
}
\seealso{
\code{\link[=set_methylCIPHER_path]{set_methylCIPHER_path()}}, \code{\link[=get_methylCIPHER_default_path]{get_methylCIPHER_default_path()}}
}
