% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{param_template}
\alias{param_template}
\title{Inherit parameters for functions}
\usage{
param_template(DNAm, pheno, imputation)
}
\arguments{
\item{DNAm}{A matrix of methylation beta values with samples as rows and CpGs as columns. Must include row names (samples) and unique column names (CpGs).}

\item{pheno}{A data.frame containing sample phenotype data, with samples as rows. Must include the following columns with no missing values:
\itemize{
  \item \code{Sample_ID}: Character vector of sample IDs to match with `DNAm`.
}}

\item{imputation}{Logical value that will allows you to perform (T)/ skip (F) imputation of mean values for missing CpGs. Warning: when imputation = F if there are missing CpGs, it will automatically ignore these CpGs during calculation, making the clock values less accurate.}
}
\value{
If you added the `pheno` input (preferred) the function appends a column with the clock calculation and returns the data.frame. Otherwise, it will return a data.frame of calculated clock values.
}
\description{
Template function to inherit parameters for other functions.
}
\keyword{internal}
