% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNAcol.R
\name{removeNAcol}
\alias{removeNAcol}
\title{Removal of empty CpG columns}
\usage{
removeNAcol(x)
}
\arguments{
\item{x}{A dataframe of CpG Betas with CpGs of all NA values}
}
\value{
Dataframe with NA columns removed
}
\description{
Removal of empty CpG columns
}
