% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylCIPHER_path.R
\name{set_methylCIPHER_path}
\alias{set_methylCIPHER_path}
\title{Set Custom Path for methylCIPHER Data}
\usage{
set_methylCIPHER_path(path)
}
\arguments{
\item{path}{A string specifying the custom path to the methylCIPHER data directory.}
}
\description{
This function allows the user to set a custom directory path for storing
methylCIPHER data, which is required for SystemsAge and PCClocks. The specified
path is stored in R's options under "methylCIPHER.path".
}
\examples{
# Returns the default path
get_methylCIPHER_path()
# Sets path to current directory
set_methylCIPHER_path(".")
# Returns the new path
get_methylCIPHER_path()
}
\seealso{
\code{\link[=set_methylCIPHER_path]{set_methylCIPHER_path()}}, \code{\link[=get_methylCIPHER_default_path]{get_methylCIPHER_default_path()}}
}
